SHL-93; Decoupling the script command from the Abstract Shell
