+++ src/main/java/org/springframework/shell/core/AbstractShell.java
@@ -15,40 +15,18 @@
  */
 package org.springframework.shell.core;
 
-import static org.springframework.shell.support.util.OsUtils.LINE_SEPARATOR;
-
-import java.io.BufferedInputStream;
-import java.io.BufferedReader;
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.URL;
-import java.text.DateFormat;
-import java.util.Collection;
-import java.util.Date;
-import java.util.Locale;
-import java.util.Map.Entry;
-import java.util.Set;
-import java.util.TreeSet;
-import java.util.logging.Level;
-import java.util.logging.Logger;
-
 import org.springframework.shell.commands.support.CommentDefinition;
-import org.springframework.shell.core.annotation.CliCommand;
-import org.springframework.shell.core.annotation.CliOption;
 import org.springframework.shell.event.AbstractShellStatusPublisher;
 import org.springframework.shell.event.ParseResult;
 import org.springframework.shell.event.ShellStatus;
 import org.springframework.shell.event.ShellStatus.Status;
 import org.springframework.shell.support.logging.HandlerUtils;
-import org.springframework.shell.support.util.IOUtils;
-import org.springframework.shell.support.util.MathUtils;
 import org.springframework.shell.support.util.VersionUtils;
 import org.springframework.util.Assert;
-import org.springframework.util.StringUtils;
+
+import java.io.File;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 /**
  * Provides a base {@link Shell} implementation.
@@ -73,96 +51,11 @@ public abstract class AbstractShell extends AbstractShellStatusPublisher impleme
 
 	protected ExitShellRequest exitShellRequest;
 
-	/**
-	 * Returns any classpath resources with the given path
-	 *
-	 * @param path the path for which to search (never null)
-	 * @return <code>null</code> if the search can't be performed
-	 * @since 1.2.0
-	 */
-	protected abstract Collection<URL> findResources(String path);
-
 	protected abstract String getHomeAsString();
 
 	protected abstract ExecutionStrategy getExecutionStrategy();
 
 	protected abstract Parser getParser();
-
-	@CliCommand(value = { "script" }, help = "Parses the specified resource file and executes its commands")
-	public void script(
-		@CliOption(key = { "", "file" }, help = "The file to locate and execute", mandatory = true) final File script,
-		@CliOption(key = "lineNumbers", mandatory = false, specifiedDefaultValue = "true", unspecifiedDefaultValue = "false", help = "Display line numbers when executing the script") final boolean lineNumbers) {
-
-		Assert.notNull(script, "Script file to parse is required");
-		double startedNanoseconds = System.nanoTime();
-		final InputStream inputStream = openScript(script);
-
-		BufferedReader in = null;
-		try {
-			in = new BufferedReader(new InputStreamReader(inputStream));
-			String line;
-			int i = 0;
-			while ((line = in.readLine()) != null) {
-				i++;
-				if (lineNumbers) {
-					logger.fine("Line " + i + ": " + line);
-				} else {
-					logger.fine(line);
-				}
-				if (!"".equals(line.trim())) {
-					boolean success = executeScriptLine(line);
-					if (success && ((line.trim().startsWith("q") || line.trim().startsWith("ex")))) {
-						break;
-					} else if (!success) {
-						// Abort script processing, given something went wrong
-						throw new IllegalStateException("Script execution aborted");
-					}
-				}
-			}
-		} catch (IOException e) {
-			throw new IllegalStateException(e);
-		} finally {
-			IOUtils.closeQuietly(inputStream, in);
-			double executionDurationInSeconds = (System.nanoTime() - startedNanoseconds) / 1000000000D;
-			logger.fine("Script required " + MathUtils.round(executionDurationInSeconds, 3) + " seconds to execute");
-		}
-	}
-
-	/**
-	 * Opens the given script for reading
-	 *
-	 * @param script the script to read (required)
-	 * @return a non-<code>null</code> input stream
-	 */
-	private InputStream openScript(final File script) {
-		try {
-			return new BufferedInputStream(new FileInputStream(script));
-		} catch (final FileNotFoundException fnfe) {
-			// Try to find the script via the classloader
-			final Collection<URL> urls = findResources(script.getName());
-
-			// Handle search failure
-			Assert.notNull(urls, "Unexpected error looking for '" + script.getName() + "'");
-
-			// Handle the search being OK but the file simply not being present
-			Assert.notEmpty(urls, "Script '" + script + "' not found on disk or in classpath");
-			Assert.isTrue(urls.size() == 1, "More than one '" + script + "' was found in the classpath; unable to continue");
-			try {
-				return urls.iterator().next().openStream();
-			} catch (IOException e) {
-				throw new IllegalStateException(e);
-			}
-		}
-	}
-
-	/**
-	 * Execute the single line from a script.
-	 * <p>
-	 * This method can be overridden by sub-classes to pre-process script lines.
-	 */
-	protected boolean executeScriptLine(final String line) {
-		return executeCommand(line);
-	}
 
 	public boolean executeCommand(String line) {
 		// Another command was attempted
